Refactor : 서버 라우터 분리

+++ server/src/routes/album.ts
@@ -1,6 +1,9 @@
 import { CustomRoute, METHOD } from '../types';
 import callSpotifyApi from '../utils/callSpotifyApi';
-const baseUrl = process.env.BASE_URL;
+import { BASE_URL_API } from '../config';
+import StatusError from '../errors/statusError';
+import { errorMessages } from '..';
+
 const albumRoute: CustomRoute[] = [
     {
         method: METHOD.GET,
@@ -9,11 +12,20 @@ const albumRoute: CustomRoute[] = [
             try {
                 const offset = query.cursor;
                 const accessToken = cookies.access_token;
-                const data = await callSpotifyApi(`${baseUrl}/v1/me/albums?limit=20&offset=${offset}`, accessToken);
-                const newData: any = data?.items.filter((a: any) => a !== null); // 알 수 없는 이유로 내부에 null 요소가 있어서 데이터 가공
-                return res.json({ ...data, items: newData });
+                if (!accessToken) {
+                    return res.status(401).json({ message: '엑세스 토큰이 존재하지 않습니다' });
+                }
+                const data = await callSpotifyApi(`${BASE_URL_API}/v1/me/albums?limit=20&offset=${offset}`, {
+                    token: accessToken,
+                });
+                const filterData = data?.items.filter((a: { album: { name: string } }) => a !== null); // 알 수 없는 이유로 내부에 null 요소가 있어서 데이터 가공
+                return res.json({ ...data, items: filterData });
             } catch (error) {
-                return res.status(500).json({ error: '서버 에러입니다.' });
+                if (error instanceof StatusError) {
+                    return res.status(error.status).json({ message: errorMessages[error.status] });
+                } else {
+                    return res.status(500).json({ message: errorMessages[500] });
+                }
             }
         },
     },
@@ -24,14 +36,21 @@ const albumRoute: CustomRoute[] = [
             try {
                 const ids = body.ids;
                 const accessToken = cookies.access_token;
-                const data = await callSpotifyApi(`${baseUrl}/v1/me/albums`, accessToken, {
+                if (!accessToken) {
+                    return res.status(401).json({ message: '엑세스 토큰이 존재하지 않습니다' });
+                }
+                await callSpotifyApi(`${BASE_URL_API}/v1/me/albums`, {
                     method: 'PUT',
+                    token: accessToken,
                     data: { ids },
                 });
-
-                return res.json({ message: '저장 완료' });
+                return res.json({ message: '앨범을 보관함에 저장하였습니다' });
             } catch (error) {
-                return res.status(500).json({ error: '서버 에러입니다.' });
+                if (error instanceof StatusError) {
+                    return res.status(error.status).json({ message: errorMessages[error.status] });
+                } else {
+                    return res.status(500).json({ message: errorMessages[500] });
+                }
             }
         },
     },
@@ -42,14 +61,21 @@ const albumRoute: CustomRoute[] = [
             try {
                 const ids = body.ids;
                 const accessToken = cookies.access_token;
-                const data = await callSpotifyApi(`${baseUrl}/v1/me/albums`, accessToken, {
+                if (!accessToken) {
+                    return res.status(401).json({ message: '엑세스 토큰이 존재하지 않습니다' });
+                }
+                await callSpotifyApi(`${BASE_URL_API}/v1/me/albums`, {
                     method: 'DELETE',
+                    token: accessToken,
                     data: { ids },
                 });
-
-                return res.json({ message: '삭제 완료' });
+                return res.json({ message: '앨범을 보관함에서 삭제하였습니다' });
             } catch (error) {
-                return res.status(500).json({ error: '서버 에러입니다.' });
+                if (error instanceof StatusError) {
+                    return res.status(error.status).json({ message: errorMessages[error.status] });
+                } else {
+                    return res.status(500).json({ message: errorMessages[500] });
+                }
             }
         },
     },
@@ -60,10 +86,19 @@ const albumRoute: CustomRoute[] = [
             try {
                 const albumId = query.id;
                 const accessToken = cookies.access_token;
-                const data = await callSpotifyApi(`${baseUrl}/v1/me/albums/contains?ids=${albumId}`, accessToken);
-                return res.json(data);
+                if (!accessToken) {
+                    return res.status(401).json({ message: '엑세스 토큰이 존재하지 않습니다' });
+                }
+                const data = await callSpotifyApi(`${BASE_URL_API}/v1/me/albums/contains?ids=${albumId}`, {
+                    token: accessToken,
+                });
+                return res.json({ setStatus: data[0] });
             } catch (error) {
-                return res.status(500).json({ error: '서버 에러입니다.' });
+                if (error instanceof StatusError) {
+                    return res.status(error.status).json({ message: errorMessages[error.status] });
+                } else {
+                    return res.status(500).json({ message: errorMessages[500] });
+                }
             }
         },
     },
